Fix compile error with testnet

+++ src/bigbang/core.cpp
@@ -881,7 +881,7 @@ Errno CCoreProtocol::VerifyBlockTx(const CTransaction& tx, const CTxContxt& txCo
     }
 
     if (destIn.IsTemplate() && destIn.GetTemplateId().GetType() == TEMPLATE_DEXMATCH
-        && nForkHeight < MATCH_VERIFY_ERROR_HEIGHT)
+        && nForkHeight < (int)MATCH_VERIFY_ERROR_HEIGHT)
     {
         nForkHeight -= 1;
     }
